refactor: remove lite text read only editor
