+++ apps/space/core/components/editor/index.ts
@@ -1,5 +1,4 @@
 export * from "./embeds";
 export * from "./lite-text-editor";
-export * from "./lite-text-read-only-editor";
 export * from "./rich-text-editor";
 export * from "./toolbar";
