+++ apps/web/core/components/editor/lite-text/index.ts
@@ -1,3 +1,2 @@
 export * from "./editor";
-export * from "./read-only-editor";
 export * from "./toolbar";
