+++ packages/editor/src/ce/extensions/core/index.ts
@@ -1,2 +1,1 @@
 export * from "./extensions";
-export * from "./read-only-extensions";
