--- packages/editor/src/ce/extensions/core/read-only-extensions.ts
@@ -1,15 +1,0 @@
-import type { Extensions } from "@tiptap/core";
-// types
-import type { IReadOnlyEditorProps } from "@/types";
-
-export type TCoreReadOnlyEditorAdditionalExtensionsProps = Pick<
-  IReadOnlyEditorProps,
-  "disabledExtensions" | "flaggedExtensions"
->;
-
-export const CoreReadOnlyEditorAdditionalExtensions = (
-  props: TCoreReadOnlyEditorAdditionalExtensionsProps
-): Extensions => {
-  const {} = props;
-  return [];
-};
