+++ packages/editor/src/core/components/editors/index.ts
@@ -4,4 +4,3 @@ export * from "./rich-text";
 export * from "./editor-container";
 export * from "./editor-content";
 export * from "./editor-wrapper";
-export * from "./read-only-editor-wrapper";
