+++ packages/editor/src/core/components/editors/lite-text/index.ts
@@ -1,2 +1,1 @@
 export * from "./editor";
-export * from "./read-only-editor";
