--- packages/editor/src/core/components/editors/lite-text/read-only-editor.tsx
@@ -1,13 +1,0 @@
-import { forwardRef } from "react";
-// components
-import { ReadOnlyEditorWrapper } from "@/components/editors";
-// types
-import { EditorReadOnlyRefApi, ILiteTextReadOnlyEditorProps } from "@/types";
-
-const LiteTextReadOnlyEditorWithRef = forwardRef<EditorReadOnlyRefApi, ILiteTextReadOnlyEditorProps>((props, ref) => (
-  <ReadOnlyEditorWrapper {...props} forwardedRef={ref as React.MutableRefObject<EditorReadOnlyRefApi | null>} />
-));
-
-LiteTextReadOnlyEditorWithRef.displayName = "LiteReadOnlyEditorWithRef";
-
-export { LiteTextReadOnlyEditorWithRef };
