+++ packages/editor/src/core/extensions/callout/index.ts
@@ -1,3 +1,1 @@
-export * from "./block";
 export * from "./extension";
-export * from "./read-only-extension";
