--- packages/editor/src/core/extensions/read-only-extensions.ts
@@ -1,102 +1,0 @@
-import { Extensions } from "@tiptap/core";
-import CharacterCount from "@tiptap/extension-character-count";
-import TaskItem from "@tiptap/extension-task-item";
-import TaskList from "@tiptap/extension-task-list";
-import TextStyle from "@tiptap/extension-text-style";
-import TiptapUnderline from "@tiptap/extension-underline";
-import { Markdown } from "tiptap-markdown";
-// extensions
-import {
-  CustomQuoteExtension,
-  CustomHorizontalRule,
-  CustomLinkExtension,
-  CustomTypographyExtension,
-  CustomCodeBlockExtension,
-  CustomCodeInlineExtension,
-  TableHeader,
-  TableCell,
-  TableRow,
-  Table,
-  CustomMentionExtension,
-  CustomTextAlignExtension,
-  CustomCalloutReadOnlyExtension,
-  CustomColorExtension,
-  UtilityExtension,
-  ImageExtension,
-} from "@/extensions";
-// plane editor extensions
-import { CoreReadOnlyEditorAdditionalExtensions } from "@/plane-editor/extensions";
-// types
-import type { IReadOnlyEditorProps } from "@/types";
-// local imports
-import { CustomImageExtension } from "./custom-image/extension";
-import { EmojiExtension } from "./emoji/extension";
-import { CustomStarterKitExtension } from "./starter-kit";
-
-type Props = Pick<IReadOnlyEditorProps, "disabledExtensions" | "flaggedExtensions" | "fileHandler" | "mentionHandler">;
-
-export const CoreReadOnlyEditorExtensions = (props: Props): Extensions => {
-  const { disabledExtensions, fileHandler, flaggedExtensions, mentionHandler } = props;
-
-  const extensions = [
-    CustomStarterKitExtension({
-      enableHistory: false,
-    }),
-    EmojiExtension,
-    CustomQuoteExtension,
-    CustomHorizontalRule,
-    CustomLinkExtension,
-    CustomTypographyExtension,
-    TiptapUnderline,
-    TextStyle,
-    TaskList.configure({
-      HTMLAttributes: {
-        class: "not-prose pl-2 space-y-2",
-      },
-    }),
-    TaskItem.configure({
-      HTMLAttributes: {
-        class: "relative pointer-events-none",
-      },
-      nested: true,
-    }),
-    CustomCodeBlockExtension,
-    CustomCodeInlineExtension,
-    Markdown.configure({
-      html: true,
-      transformCopiedText: false,
-    }),
-    Table,
-    TableHeader,
-    TableCell,
-    TableRow,
-    CustomMentionExtension(mentionHandler),
-    CharacterCount,
-    CustomColorExtension,
-    CustomTextAlignExtension,
-    CustomCalloutReadOnlyExtension,
-    UtilityExtension({
-      disabledExtensions,
-      fileHandler,
-      isEditable: false,
-    }),
-    ...CoreReadOnlyEditorAdditionalExtensions({
-      disabledExtensions,
-      flaggedExtensions,
-    }),
-  ];
-
-  if (!disabledExtensions.includes("image")) {
-    extensions.push(
-      ImageExtension({
-        fileHandler,
-      }),
-      CustomImageExtension({
-        fileHandler,
-        isEditable: false,
-      })
-    );
-  }
-
-  return extensions;
-};
