+++ packages/editor/src/core/hooks/use-editor.ts
@@ -1,23 +1,16 @@
-import { DOMSerializer } from "@tiptap/pm/model";
 import { useEditorState, useEditor as useTiptapEditor } from "@tiptap/react";
 import { useImperativeHandle, useEffect } from "react";
-import * as Y from "yjs";
-// components
-import { getEditorMenuItems } from "@/components/menus";
 // constants
 import { CORE_EXTENSIONS } from "@/constants/extension";
 // extensions
 import { CoreEditorExtensions } from "@/extensions";
 // helpers
-import { getParagraphCount } from "@/helpers/common";
 import { getEditorRefHelpers } from "@/helpers/editor-ref";
 import { getExtensionStorage } from "@/helpers/get-extension-storage";
-import { insertContentAtSavedSelection } from "@/helpers/insert-content-at-cursor-position";
-import { scrollToNodeViaDOMCoordinates } from "@/helpers/scroll-to-node";
 // props
 import { CoreEditorProps } from "@/props";
 // types
-import type { TEditorCommands, TEditorHookProps } from "@/types";
+import type { TEditorHookProps } from "@/types";
 
 export const useEditor = (props: TEditorHookProps) => {
   const {
@@ -124,155 +117,7 @@ export const useEditor = (props: TEditorHookProps) => {
     onAssetChange(assets);
   }, [assetsList?.assets, onAssetChange]);
 
-  useImperativeHandle(
-    forwardedRef,
-    () => ({
-      ...getEditorRefHelpers({ editor, provider }),
-      blur: () => editor?.commands.blur(),
-      emitRealTimeUpdate: (message) => provider?.sendStateless(message),
-      executeMenuItemCommand: (props) => {
-        const { itemKey } = props;
-        const editorItems = getEditorMenuItems(editor);
-
-        const getEditorMenuItem = (itemKey: TEditorCommands) => editorItems.find((item) => item.key === itemKey);
-
-        const item = getEditorMenuItem(itemKey);
-        if (item) {
-          item.command(props);
-        } else {
-          console.warn(`No command found for item: ${itemKey}`);
-        }
-      },
-      getCurrentCursorPosition: () => editor?.state.selection.from,
-      getSelectedText: () => {
-        if (!editor) return null;
-
-        const { state } = editor;
-        const { from, to, empty } = state.selection;
-
-        if (empty) return null;
-
-        const nodesArray: string[] = [];
-        state.doc.nodesBetween(from, to, (node, _pos, parent) => {
-          if (parent === state.doc && editor) {
-            const serializer = DOMSerializer.fromSchema(editor.schema);
-            const dom = serializer.serializeNode(node);
-            const tempDiv = document.createElement("div");
-            tempDiv.appendChild(dom);
-            nodesArray.push(tempDiv.innerHTML);
-          }
-        });
-        const selection = nodesArray.join("");
-        return selection;
-      },
-      insertText: (contentHTML, insertOnNextLine) => {
-        if (!editor) return;
-        const { from, to, empty } = editor.state.selection;
-        if (empty) return;
-        if (insertOnNextLine) {
-          // move cursor to the end of the selection and insert a new line
-          editor.chain().focus().setTextSelection(to).insertContent("<br />").insertContent(contentHTML).run();
-        } else {
-          // replace selected text with the content provided
-          editor.chain().focus().deleteRange({ from, to }).insertContent(contentHTML).run();
-        }
-      },
-      isEditorReadyToDiscard: () =>
-        !!editor && getExtensionStorage(editor, CORE_EXTENSIONS.UTILITY)?.uploadInProgress === false,
-      isMenuItemActive: (props) => {
-        const { itemKey } = props;
-        const editorItems = getEditorMenuItems(editor);
-
-        const getEditorMenuItem = (itemKey: TEditorCommands) => editorItems.find((item) => item.key === itemKey);
-        const item = getEditorMenuItem(itemKey);
-        if (!item) return false;
-
-        return item.isActive(props);
-      },
-      listenToRealTimeUpdate: () => provider && { on: provider.on.bind(provider), off: provider.off.bind(provider) },
-      onDocumentInfoChange: (callback) => {
-        const handleDocumentInfoChange = () => {
-          if (!editor) return;
-          callback({
-            characters: editor ? getExtensionStorage(editor, CORE_EXTENSIONS.CHARACTER_COUNT)?.characters?.() : 0,
-            paragraphs: getParagraphCount(editor?.state),
-            words: editor ? getExtensionStorage(editor, CORE_EXTENSIONS.CHARACTER_COUNT)?.words?.() : 0,
-          });
-        };
-
-        // Subscribe to update event emitted from character count extension
-        editor?.on("update", handleDocumentInfoChange);
-        // Return a function to unsubscribe to the continuous transactions of
-        // the editor on unmounting the component that has subscribed to this
-        // method
-        return () => {
-          editor?.off("update", handleDocumentInfoChange);
-        };
-      },
-      onHeadingChange: (callback) => {
-        const handleHeadingChange = () => {
-          if (!editor) return;
-          const headings = getExtensionStorage(editor, CORE_EXTENSIONS.HEADINGS_LIST)?.headings;
-          if (headings) {
-            callback(headings);
-          }
-        };
-
-        // Subscribe to update event emitted from headers extension
-        editor?.on("update", handleHeadingChange);
-        // Return a function to unsubscribe to the continuous transactions of
-        // the editor on unmounting the component that has subscribed to this
-        // method
-        return () => {
-          editor?.off("update", handleHeadingChange);
-        };
-      },
-      onStateChange: (callback) => {
-        // Subscribe to editor state changes
-        editor?.on("transaction", callback);
-
-        // Return a function to unsubscribe to the continuous transactions of
-        // the editor on unmounting the component that has subscribed to this
-        // method
-        return () => {
-          editor?.off("transaction", callback);
-        };
-      },
-      scrollToNodeViaDOMCoordinates(behavior, pos) {
-        const resolvedPos = pos ?? editor?.state.selection.from;
-        if (!editor || !resolvedPos) return;
-        scrollToNodeViaDOMCoordinates(editor, resolvedPos, behavior);
-      },
-      setEditorValueAtCursorPosition: (content) => {
-        if (editor?.state.selection) {
-          insertContentAtSavedSelection(editor, content);
-        }
-      },
-      setFocusAtPosition: (position) => {
-        if (!editor || editor.isDestroyed) {
-          console.error("Editor reference is not available or has been destroyed.");
-          return;
-        }
-        try {
-          const docSize = editor.state.doc.content.size;
-          const safePosition = Math.max(0, Math.min(position, docSize));
-          editor
-            .chain()
-            .insertContentAt(safePosition, [{ type: CORE_EXTENSIONS.PARAGRAPH }])
-            .focus()
-            .run();
-        } catch (error) {
-          console.error("An error occurred while setting focus at position:", error);
-        }
-      },
-      setProviderDocument: (value) => {
-        const document = provider?.document;
-        if (!document) return;
-        Y.applyUpdate(document, value);
-      },
-    }),
-    [editor, provider]
-  );
+  useImperativeHandle(forwardedRef, () => getEditorRefHelpers({ editor, provider }), [editor, provider]);
 
   if (!editor) {
     return null;
