--- packages/editor/src/core/props/index.ts
@@ -1,2 +1,0 @@
-export * from "./props";
-export * from "./read-only";
