+++ packages/editor/src/core/types/config.ts
@@ -1,17 +1,14 @@
 // plane imports
 import { TWebhookConnectionQueryParams } from "@plane/types";
 
-export type TReadOnlyFileHandler = {
+export type TFileHandler = {
+  assetsUploadStatus: Record<string, number>; // blockId => progress percentage
+  cancel: () => void;
   checkIfAssetExists: (assetId: string) => Promise<boolean>;
+  delete: (assetSrc: string) => Promise<void>;
   getAssetDownloadSrc: (path: string) => Promise<string>;
   getAssetSrc: (path: string) => Promise<string>;
   restore: (assetSrc: string) => Promise<void>;
-};
-
-export type TFileHandler = TReadOnlyFileHandler & {
-  assetsUploadStatus: Record<string, number>; // blockId => progress percentage
-  cancel: () => void;
-  delete: (assetSrc: string) => Promise<void>;
   upload: (blockId: string, file: File) => Promise<string>;
   validation: {
     /**
